--- a/inst/include/cpp11/parse.hpp
+++ b/inst/include/cpp11/parse.hpp
@@ -6,7 +6,7 @@
 
 using simdjson::ondemand::json_type;
 
-inline auto bad_json_type_message(simdjson::ondemand::value element, const std::string& expected, JSON_Path& path) {
+inline auto bad_json_type_message(simdjson::ondemand::value element, const std::string& expected, const JSON_Path& path) {
     return "Cannot convert a JSON " + json_type_to_string(element) + " to " + expected + " at path " + path.path();
 }
 
@@ -26,7 +26,7 @@ inline int parse_scalar_bool(simdjson::ondemand::value element, JSON_Path& path)
     }
 }
 
-inline int parse_scalar_int(simdjson::ondemand::value element, JSON_Path& path) {
+inline int parse_scalar_int(simdjson::ondemand::value element, const JSON_Path& path) {
     switch (element.type()) {
     case json_type::number:
         // TODO this is unsafe; add handling of big integers
@@ -40,7 +40,7 @@ inline int parse_scalar_int(simdjson::ondemand::value element, JSON_Path& path) 
     }
 }
 
-inline double parse_scalar_double(simdjson::ondemand::value element, JSON_Path& path) {
+inline double parse_scalar_double(simdjson::ondemand::value element, const JSON_Path& path) {
     switch (element.type()) {
     case json_type::number:
         return static_cast<double>(element);
@@ -53,7 +53,7 @@ inline double parse_scalar_double(simdjson::ondemand::value element, JSON_Path& 
     }
 }
 
-inline SEXPREC* parse_scalar_string(simdjson::ondemand::value element, JSON_Path& path) {
+inline SEXPREC* parse_scalar_string(simdjson::ondemand::value element, const JSON_Path& path) {
     switch (element.type()) {
     case json_type::string:
         return Rf_mkChar(std::string(std::string_view(element)).c_str());