* use `std::string_view`
  * add a private member value `std::vector<std::unique_ptr<std::string>> string_view_protection;`
  * map key must point to the corresponding value in `string_view_protection`
* only search once for a key instead of twice
* remove `key_found` map by storing the info in the Column class

Without the inefficient path it is now faster than `RcppSimdJson`
```r
# A tibble: 2 × 13
  expression        min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result memory 
  <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list> <list> 
1 jsonparse      3.42ms   3.84ms      253.     238KB     0      100     0      395ms <NULL> <Rprof…
2 RcppSimdJson   4.13ms    4.8ms      204.    3.59MB     2.06    99     1      485ms <NULL> <Rprof…
```